архив ЗЛП без льготы

+++ processor/src/core/java/ru/axetta/ecafe/processor/core/persistence/service/clients/ClientMigrationHistoryService.java
@@ -96,7 +96,11 @@ public class ClientMigrationHistoryService {
                         }
                     }
                     Client client = DAOUtils.findClient(session, clientMigration.getClient().getIdOfClient());
-                    ClientManager.deleteDiscount(client, session);
+                    if (ClientManager.atLeastOneDiscountEligibleToDelete(client)) {
+                        ClientManager.deleteDiscount(client, session);
+                    } else {
+                        ClientManager.archiveApplicationForFoodWithoutDiscount(client, session);
+                    }
                     transaction.commit();
                     transaction = null;
                     counter++;
